fix(services): extract shared build_output_files into services/output_plan (Fixes #352)
dry_run_service.rs and install_service.rs had duplicated copies of
PlannedOutputFile, build_output_files, push_base_plans,
push_base_output_files, and push_droid_output_files — the entire
output-plan-building phase was copied. When a new plugin is added,
both copies must be updated, guaranteeing drift.

Extract into a shared sdk/src/services/output_plan.rs module. Both
services now import the same build_output_files. The only difference
between install and dry-run is the write phase:
- install_service calls write_output_files (actual disk writes)
- dry_run_service skips it entirely (this IS the 'mock')

Files changed:
  sdk/src/services/output_plan.rs          | +141  (new shared module)
  sdk/src/services/dry_run_service.rs      | -121  (remove duplicates)
  sdk/src/services/install_service.rs      | -209  (remove duplicates)
  sdk/src/services/mod.rs                  | +1    (pub mod output_plan)
