+++ sdk/src/services/mod.rs
@@ -3,4 +3,5 @@ pub mod command_diagnostics_service;
 pub mod common;
 pub mod dry_run_service;
 pub mod install_service;
+pub mod output_plan;
 pub mod prompt_service;
